Measures

A class and utility methods for numeric values with units

Usage

measure(x = double(), units = attr(x, "units"))

# S4 method for class 'measure'
units(x)

# S4 method for class 'measure'
units(x) <- value

Arguments

x

a numeric vector, or a measure, difftime, lubridate::duration, lubridate::interval, or lubridate::period object

units, value

a character value. See Details

Value

An object of class measure

Methods

If x is a lubridate Duration, Interval or Period object, it will be converted to difftime before running the default method.

If x is already a measure convert_units() will be called.

Units

Units are specified in the following format:

"multiplier prefix1|unit1^exponent1 prefix2|unit2^exponent2 ...".

The multiplier, prefixes and exponents are optional. For example, measure(x, "10 c|m^2 s^-1") will add units of ten centimeters per second, and measure(x, "m") simply of meters. Prefixes can be any of the SI prefixes:

Q, R, Y, Z, E, P, T, G, M, k, h, da, d, c, m, µ, (or micro), n, p, f, a, z, y, r, or q.

Units have no restriction, although difftime units are converted to s, min, hour, day and week.

See also