Reorganize Measurements to Two Columns

step_measure_output_long creates a specification of a recipe step that converts measures to a format with columns for the measurement and the corresponding location (i.e., "long" format).

Usage

step_measure_output_long(
  recipe,
  values_to = ".measure",
  location_to = ".location",
  role = "predictor",
  trained = FALSE,
  skip = FALSE,
  id = rand_id("measure_output_long")
)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

values_to

A single character string for the column containing the analytical maesurement.

location_to

A single character string for the column containing the location of the measurement (e.g. wavenumber or index).

role

Not used by this step since no new variables are created.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

skip

A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A character string that is unique to this step to identify it.

Details

This step is designed convert analytical measurements from their internal data structure to a two column format.

See also

Examples

library(dplyr)

data(glucose_bioreactors)
bioreactors_small$batch_sample <- NULL

small_tr <- bioreactors_small[  1:200,]
small_te <- bioreactors_small[201:210,]

small_rec <-
  recipe(glucose ~ . , data = small_tr) %>%
  update_role(batch_id, day, new_role = "id columns") %>%
  step_measure_input_wide(`400`:`3050`) %>%
  prep()

# Before reformatting:

small_rec %>% bake(new_data = small_te)
#> # A tibble: 10 × 4
#>    batch_id   day glucose .measures           
#>    <fct>    <int>   <dbl> <list>              
#>  1 S_15         5    2.04 <tibble [2,651 × 2]>
#>  2 S_15         6    5.56 <tibble [2,651 × 2]>
#>  3 S_15         7    4.65 <tibble [2,651 × 2]>
#>  4 S_15         8    9.91 <tibble [2,651 × 2]>
#>  5 S_15         9    4.96 <tibble [2,651 × 2]>
#>  6 S_15        10    6.78 <tibble [2,651 × 2]>
#>  7 S_15        11    6.72 <tibble [2,651 × 2]>
#>  8 S_15        12    4.69 <tibble [2,651 × 2]>
#>  9 S_15        13    6.30 <tibble [2,651 × 2]>
#> 10 S_15        14    3.10 <tibble [2,651 × 2]>

# After reformatting:

output_rec <-
  small_rec %>%
  step_measure_output_long() %>%
  prep()

output_rec %>% bake(new_data = small_te)
#> Warning: ! There was 1 column that was a factor when the recipe was prepped:
#> • `batch_id`
#> ℹ This may cause errors when processing new data.
#> # A tibble: 26,510 × 5
#>    batch_id   day glucose .measure .location
#>    <fct>    <int>   <dbl>    <dbl>     <int>
#>  1 S_15         5    2.04  760094.         1
#>  2 S_15         5    2.04  779053.         2
#>  3 S_15         5    2.04  797154.         3
#>  4 S_15         5    2.04  817226.         4
#>  5 S_15         5    2.04  832725.         5
#>  6 S_15         5    2.04  840075.         6
#>  7 S_15         5    2.04  841721.         7
#>  8 S_15         5    2.04  832112.         8
#>  9 S_15         5    2.04  819494.         9
#> 10 S_15         5    2.04  799601.        10
#> # ℹ 26,500 more rows